ref(settings): Move settings button to toolbar
Move settings button component to toolbar folder to keep the general
folder structure
