--- angular-frontend/src/app/settings/settings-button.component.css
@@ -1,36 +1,0 @@
-/* Setting Menu Button Styles */
-button {
-  color: var(--mat-sys-on-surface);
-  transition: all 0.2s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 8px;
-}
-
-button:hover {
-  color: var(--mat-sys-primary);
-  transform: scale(1.05);
-}
-
-button:active {
-  transform: scale(0.95);
-}
-
-.mat-icon {
-  font-size: var(--font-size-large);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 !important;
-  vertical-align: middle;
-}
-
-/* Dark theme support */
-:host-context(.dark-theme) button {
-  color: var(--mat-sys-on-surface);
-}
-
-:host-context(.dark-theme) button:hover {
-  color: var(--mat-sys-primary);
-}
