+++ angular-frontend/src/app/toolbar/settings-button.component.html
